Work-in-progress:  adding new routes
Working through the "Core Concepts and Reading Resources" portion, and
the examples provided there.  The new route for ingredients resulted in
a server-side error, which was intentional on the part of the instructor!

A graph was detected in the object graph when trying to do JSON-manipulation,
perhaps?  And part of the lesson is intended to explain the problems with
directly exposing entities through a REST-ish interface?